+++ src/container/SignUpContainer.js
@@ -1,11 +1,8 @@
 import React from "react";
 
 import styled from "styled-components";
-import { setColor } from "../style/SetStyle";
 import { Div } from "../style/LayoutStyle";
-import { Input } from "../style/InputStyle";
 import { Button } from "../style/ButtonStyle";
-import { P } from "../style/TextStyle";
 import { Img } from "../style/ImgStyle";
 import { Section } from "../style/LayoutStyle";
 
@@ -13,8 +10,7 @@ import InputItem from "../component/InputItem";
 import MainLogo from "../img/HeaderLogo.svg";
 import KakaoLoginBtn from "../img/kakaoLoginMediumWide.svg";
 
-import { useClick } from "../hook/useClick";
-import AgreementContainer from "./AgreementContainer";
+import TermsServiceContainer from "./TermsServiceContainer";
 
 const dummyIdData = {
   id: {
@@ -59,11 +55,6 @@ const dummyPWData = {
   },
 };
 
-const BorderStyleBtn = styled(Button)`
-  border: 1px solid ${setColor("black")};
-  border-radius: 4px;
-`;
-
 const BorderStyleNoneBtn = styled(Button)`
   border-style: none;
   border-radius: 4px;
@@ -74,20 +65,6 @@ const KakaoLoginStyleBtn = styled(Img)`
 `;
 
 const SignUpContainer = () => {
-  // const [isAllChecked, setIsAllChecked] = useClick(false);
-
-  // const agreeAllBtnClickEvent = () => {
-  //   if (!isAllChecked) {₩
-  //     setTermsServiceChecked(!termsServiceChecked);
-  //     setPrivacyPolicyChecked(!privacyPolicyChecked);
-  //     setIsAllChecked(true);
-  //   } else {
-  //     setTermsServiceChecked(!termsServiceChecked);
-  //     setPrivacyPolicyChecked(!privacyPolicyChecked);
-  //     setIsAllChecked(false);
-  //   }
-  // };
-
   return (
     <>
       <Section $padding="50px 0" $margin="70px 0 0 0" $width="100vw" $flex="h_center_center">
@@ -99,7 +76,7 @@ const SignUpContainer = () => {
           <InputItem {...{ dummyInputData: dummyNameData }}></InputItem>
           <InputItem {...{ dummyInputData: dummyPWData }}></InputItem>
           <Div $width="100%" $flex="v_start_start">
-            <AgreementContainer />
+            <TermsServiceContainer />
             <BorderStyleNoneBtn
               $width="100%"
               $height="50px"
